Methods marked with 'ForwardRef' are now always replaced by the matching methods
